Mayor documentación con Swagger usando XML

+++ spotify-clone-backend/Controllers/TracksController.cs
@@ -1,9 +1,10 @@
-﻿using Microsoft.AspNetCore.Authorization;
+﻿using System.Net;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using spotify_clone_backend.Models;
 using spotify_clone_backend.Repositories;
 using System;
-using System.Collections.Generic;
+using Microsoft.AspNetCore.Http;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -13,6 +14,7 @@ namespace spotify_clone_backend.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Produces("application/json")]
     public class TracksController : ControllerBase
     {
         private ITrackRepository _repository;
@@ -21,7 +23,10 @@ namespace spotify_clone_backend.Controllers
         {
             _repository = repository;
         }
-
+        /// <summary>Get all tracks</summary>
+        /// <response code="200">If get all songs correctly</response>
+        /// <response code="401">If you not authenticate</response>
+        /// <response code="500">If the server has a problem</response>
         [HttpGet]
         [Authorize]
         public IActionResult Get()
@@ -37,8 +42,58 @@ namespace spotify_clone_backend.Controllers
 
             }      
         }
+        /// <summary>Get track with Id</summary>
+        /// <response code="200">If get all songs correctly</response>
+        /// <response code="401">If you not authenticate</response>
+        /// <response code="403">If you don't have authorization for use this method</response>
+        /// <response code="500">If the server has a problem</response>
+        [HttpGet("{id}")]
+        [Authorize(Roles="Administrator")]
+        public IActionResult GetTrack(long id){
 
-        [HttpPut]
+            try{
+                var track = _repository.GetTrackWithId(id);
+                if(track == null)
+                     return NotFound("Song with Id " + id + " doesnt exist" );
+                return Ok(track);
+                
+            } catch (Exception e)
+            {
+                return StatusCode(500, e.Message);
+            }
+   
+        }
+
+        /// <summary>Upload a track</summary>
+        /// <remarks>
+        /// Sample request:
+        ///      
+        ///      POST /Tracks 
+        ///      {
+        ///        "id": 0,
+        ///        "name": "Me estas tentando",
+        ///        "album": "Me estas tentando",
+        ///        <![CDATA["cover": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQ9o_E4nq6SmI0U8Rfv0IV1t_eaTmBLbqnGzg&usqp=CAU",]]>
+        ///        "artist": {
+        ///          "id": 0,
+        ///          <![CDATA["name": "Wisin & Yandel",]]>
+        ///          <![CDATA["nickname": "Wisin & Yandel",]]>
+        ///          "nationality": "PR"
+        ///        },
+        ///        "duration": {
+        ///          "id": 0,
+        ///          "start": 0,
+        ///          "end": 333
+        ///        },
+        ///        "url": "../assets/songs/track-8.mp3"
+        ///      }
+        /// </remarks>
+        /// <param name="track">Json of Track with Artist and Duration</param>
+        /// <response code="200">If the song has correct uploaded</response>
+        /// <response code="401">If you not authenticate</response>
+        /// <response code="403">If you don't have authorization for use this method</response>
+        /// <response code="500">If the server has a problem</response>
+        [HttpPost]
         [Authorize(Roles ="Administrator")]
         public IActionResult UploadTrack([FromBody] Track track){
             try{
@@ -48,18 +103,32 @@ namespace spotify_clone_backend.Controllers
                 return StatusCode(500, e.Message);
             }
         }
+        
 
-        [HttpDelete("/api/Tracks/{id}")]
+        
+        /// <summary>
+        /// Delete a specific track.
+        /// </summary>
+        /// <param name="id">Specify Id for the song you want to delete</param>
+        /// <response code="200">If the song has correct deleted</response>
+        /// <response code="401">If you not authenticate</response>
+        /// <response code="403">If you don't have authorization for use this method</response>
+        /// <response code="404">If the song with the id you sent does not exist</response>
+        /// <response code="500">If the server has a problem</response>
+        [HttpDelete("{id}")]
         [Authorize(Roles = "Administrator")]
         public IActionResult DeleteTrack(long id){
             try{
                 var track = _repository.GetTrackWithId(id);
+                if(track == null)
+                    return NotFound("Song with Id " + id + " doesnt exist" );
                 _repository.DeleteTrack(track);
                 return Ok();
             }catch(Exception e){
                 return StatusCode(500, e.Message);
             }
         }
+         
         
     }
 }
